This sould work, lets just test it first

+++ auto_setup.sh
@@ -19,26 +19,72 @@ sudo apt install build-essentials -y
 sudo apt install gcc -y
 sudo apt install gcc-13 -y
 sudo apt install gcc-14 -y
-sudo apt update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-13 1
-sudo apt update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-14 2
+sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-13 1
+sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-14 2
+
+
+echo "______Folder Setup________"
+
+cd ~/Documents
+
+DIRECTORY=~/Documents/GitHub
+if [ ! -d "$DIRECTORY"]; then
+    mkdir GitHub
+else 
+    echo "Github folder already exists"
+fi
+
+DIRECTORY=~/Documents/Local
+if [ ! -d "$DIRECTORY"]; then
+    mkdir Local
+else
+    echo "Local folder already exists"
+fi
+
+
+
+echo "______KiCad Install_______"
+
+cd ~/Documents
+DIRECTORY=~/Documents/KiCad
+if [ ! -d "$DIRECTORY"]; then
+    mkdir KiCad
+else
+    echo "KiCad folder already created"
+fi
+
+if ! [ -x "$(command -v kicad)" ]; then
+    cd ~
+    yes '' | sudo add-apt-repository ppa:kicad/kicad-8.0-releases
+    sudo apt update
+    yes | sudo apt install kicad
+else
+    echo "KiCad is already installed"
+fi
+
+
 
 echo "______CMake Install_______"
 
-sudo apt-get update
-sudo apt-get install ca-certificates gpg wget
+if ! [ -x "$(command -v cmake)" ]; then
+    sudo apt-get update
+    sudo apt-get install ca-certificates gpg wget
 
-test -f /usr/share/doc/kitware-archive-keyring/copyright ||
-wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /usr/share/keyrings/kitware-archive-keyring.gpg >/dev/null
+    test -f /usr/share/doc/kitware-archive-keyring/copyright ||
+    wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /usr/share/keyrings/kitware-archive-keyring.gpg >/dev/null
 
-echo 'deb [signed-by=/usr/share/keyrings/kitware-archive-keyring.gpg] https://apt.kitware.com/ubuntu/ noble main' | sudo tee /etc/apt/sources.list.d/kitware.list >/dev/null
-sudo apt-get update
+    echo 'deb [signed-by=/usr/share/keyrings/kitware-archive-keyring.gpg] https://apt.kitware.com/ubuntu/ noble main' | sudo tee /etc/apt/sources.list.d/kitware.list >/dev/null
+    sudo apt-get update
 
-test -f /usr/share/doc/kitware-archive-keyring/copyright ||
-sudo rm /usr/share/keyrings/kitware-archive-keyring.gpg
+    test -f /usr/share/doc/kitware-archive-keyring/copyright ||
+    sudo rm /usr/share/keyrings/kitware-archive-keyring.gpg
 
-sudo apt-get install kitware-archive-keyring
+    sudo apt-get install kitware-archive-keyring
 
-yes | sudo apt-get install cmake
+    yes | sudo apt-get install cmake
+else
+    echo "cmake is already installed"
+fi
 
 
 
@@ -52,7 +98,7 @@ if [ ! -d "$DIRECTORY" ]; then
     wget "https://download.jetbrains.com/python/pycharm-professional-2024.3.1.1.tar.gz"
     sudo tar xzf pycharm-*.tar.gz -C /opt/
     sudo rm pycharm-*.tar.gz
-    echo $"\nexport PATH="/opt/pycharm-2024.3.1.1/bin:$PATH""
+    echo $"\nexport PATH="/opt/pycharm-2024.3.1.1/bin:$PATH"" >> ~/.bashrc
 else
     echo "Pycharm is already installed ${DIRECTORY}"
 fi
@@ -69,32 +115,45 @@ if [ ! -d "$DIRECTORY" ]; then
     wget "https://download.jetbrains.com/cpp/CLion-2024.3.1.1.tar.gz"
     sudo tar xvzf CLion-*.tar.gz -C /opt/
     sudo rm CLion-*.tar.gz
-    echo $"\nexport PATH="/opt/clion-2024.3.1.1/bin:$PATH""
+    echo $"\nexport PATH="/opt/clion-2024.3.1.1/bin:$PATH"" >> ~/.bashrc
 else
     echo "CLion is already installed ${DIRECTORY}"
 fi
 
 
-echo "______VS Code Setup_______"
 
-cd ~/Downloads
-wget "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64" -O vscode.deb
-sudo apt install ./vscode.deb
-sudo rm vscode.deb
+echo "______VS Code Setup________"
+
+if ! [ -x "$(command -v code)" ]; then
+    cd ~/Downloads
+    wget "https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64" -O vscode.deb
+    sudo apt install ./vscode.deb
+    sudo rm vscode.deb
+else
+    echo "Visual Studio Code is already installed"
+fi
 
 
 
-echo "______Conda Install_______"
+echo "______Conda Install________"
 
 cd ~
-yes | sudo apt install libgl1-mesa-glx libegl1-mesa libxrandr2 libxrandr2 libxss1 libxcursor1 libxcomposite1 libasound2 libxi6 libxtst6
-curl -O https://repo.anaconda.com/archive/Anaconda3-2024.10-1-Linux-x86_64.sh
-bash ~/Anaconda3-2024.10-1-Linux-x86_64.sh -b
-conda config --set auto_activate_base True
+DIRECTORY=/anaconda3
+
+if [ ! -d "$DIRECTORY" ]; then
+    yes | sudo apt install libgl1-mesa-glx libegl1-mesa libxrandr2 libxrandr2 libxss1 libxcursor1 libxcomposite1 libasound2 libxi6 libxtst6
+    curl -O https://repo.anaconda.com/archive/Anaconda3-2024.10-1-Linux-x86_64.sh
+    bash ~/Anaconda3-2024.10-1-Linux-x86_64.sh -b
+    source ~/anaconda3/bin/activate
+    conda init --all
+    conda config --set auto_activate_base True
+else
+    echo "Conda is allready installed"
+fi
 
 
 
-echo "______Discord Setup_______"
+echo "______Discord Setup________"
 
 cd ~/Downloads
 wget 'https://discord.com/api/download?platform=linux&format=deb' -O discord.deb
@@ -103,18 +162,27 @@ sudo rm discord.deb
 
 
 
-echo "______TeamSpeak3 Setup____"
+echo "______TeamSpeak3 Setup_____"
 
 cd ~
-wget "https://files.teamspeak-services.com/releases/client/3.6.2/TeamSpeak3-Client-linux_amd64-3.6.2.run"
-chmod +x TeamSpeak3-Client-linux_amd64-3.6.2.run
-yes | sudo ./TeamSpeak3-Client-linux_amd64-3.6.2.run
+DIRECTORY=TeamSpeak3-Client-linux_amd64/
+if [ ! -d "$DIRECTORY" ]; then
+    wget "https://files.teamspeak-services.com/releases/client/3.6.2/TeamSpeak3-Client-linux_amd64-3.6.2.run"
+    chmod +x TeamSpeak3-Client-linux_amd64-3.6.2.run
+    yes | sudo ./TeamSpeak3-Client-linux_amd64-3.6.2.run
+    sudo rm TeamSpeak3-Client-linux_amd64-3.6.2.run
+else
+    echo "Teamspeak 3 client already installed"
 
 
 
-echo "______Steam Setup_________"
+echo "______Steam Setup__________"
 
-cd ~/Downloads
-wget https://cdn.fastly.steamstatic.com/client/installer/steam.deb -O steam.deb
-sudo apt install ./steam.deb
-sudo rm steam.deb
+if ! [ -x "$(command -v steam)" ]; then
+    cd ~/Downloads
+    wget https://cdn.fastly.steamstatic.com/client/installer/steam.deb -O steam.deb
+    sudo apt install ./steam.deb
+    sudo rm steam.deb
+else
+    echo "steam is already installed"
+fi
